requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.135.1
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
jiter==0.13.0
kiwisolver==1.5.0
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.9
ml_dtypes==0.5.4
moviepy==2.2.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
openai==2.26.0
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opt_einsum==3.4.0
packaging==26.0
passlib==1.7.4
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==11.3.0
proglog==0.1.12
proto-plus==1.27.1
protobuf==3.20.3
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyparsing==3.3.2
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
regex==2026.2.28
requests==2.32.5
rsa==4.9.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
SQLAlchemy==2.0.48
starlette==0.52.1
sympy==1.14.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
